Add files via upload
Provided that Test Case 11 is not a typographical error, and is done so on purpose, then it contains conditions that are not mentioned in the description of the task. The program presented here assumes that test case is so by design in order to test the coders. Further comments about this test case is contained in the corresponding sections of the program.